added citites

+++ src/Controller/CitiesController.php
@@ -15,12 +15,12 @@ class CitiesController extends AppController {
  *
  * @return void
  */
-
 	public function initialize() {
-        parent::initialize();
-        $this->loadComponent('RequestHandler');
+		parent::initialize();
+		$this->loadComponent('RequestHandler');
 		$this->response->header('Access-Control-Allow-Origin', '*');
-    }
+		$this->response->header('Access-Control-Allow-Methods', 'POST, GET, OPTIONS, PUT, DELETE');
+	}
 
 /**
  * Index method
@@ -28,11 +28,16 @@ class CitiesController extends AppController {
  * @return void
  */
 	public function index() {
-		$this->paginate = [
-			'contain' => ['Creators', 'Modifiers', 'Countries']
-		];
-		$this->set('cities', $this->paginate($this->Cities));
-        $this->set('_serialize', ['cities']);
+		if ($this->request->params['_ext']) {
+			$conditions = [
+				'fields' => ['Cities.id', 'Cities.name']
+			];
+			$this->set('cities', $this->Cities->find('all', $conditions));
+		}
+		else {
+			$this->set('cities', $this->Cities->find('all'));
+		}
+		$this->set('_serialize', ['cities']);
 	}
 
 /**
@@ -43,11 +48,17 @@ class CitiesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function view($id = null) {
-		$city = $this->Cities->get($id, [
-			'contain' => ['Creators', 'Modifiers', 'Countries', 'Customers', 'VesselOwners']
-		]);
-		$this->set('city', $city);
-        $this->set('_serialize', ['city']);
+		if ($this->request->params['_ext']) {
+			$city = $this->Cities->get($id);
+			$this->set('city', $city);
+			$this->set('_serialize', ['city']);
+		} else {
+			$city = $this->Cities->get($id, [
+				'contain' => ['Creators', 'Modifiers', 'Cities', 'Ports']
+			]);
+			$this->set('city', $city);
+			$this->set('_serialize', ['city']);
+		}
 	}
 
 /**
@@ -56,19 +67,33 @@ class CitiesController extends AppController {
  * @return void
  */
 	public function add() {
-		$city = $this->Cities->newEntity($this->request->data);
-		if ($this->request->is('post')) {
-			if ($this->Cities->save($city)) {
-				$this->Flash->success('The city has been saved.');
-				return $this->redirect(['action' => 'index']);
+		if ($this->request->params['_ext']) {
+			$city = $this->Cities->newEntity($this->request->data);
+			if ($this->Cities->save($city, ['validate' => false])) {
+				$message = 'Saved';
 			} else {
-				$this->Flash->error('The city could not be saved. Please, try again.');
+				$message = 'Error';
 			}
+			$this->set([
+				'data' => $this->request->data,
+				'message' => $message,
+				'city' => $city,
+				'_serialize' => ['message', 'city', 'data']
+			]);
+		} else {
+			$city = $this->Cities->newEntity($this->request->data);
+			if ($this->request->is('post')) {
+				if ($this->Cities->save($city)) {
+					$this->Flash->success('The city has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The city could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Cities->Creators->find('list');
+			$modifiers = $this->Cities->Modifiers->find('list');
+			$this->set(compact('city', 'creators', 'modifiers'));
 		}
-		$creators = $this->Cities->Creators->find('list');
-		$modifiers = $this->Cities->Modifiers->find('list');
-		$countries = $this->Cities->Countries->find('list');
-		$this->set(compact('city', 'creators', 'modifiers', 'countries'));
 	}
 
 /**
@@ -79,22 +104,40 @@ class CitiesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function edit($id = null) {
-		$city = $this->Cities->get($id, [
-			'contain' => []
-		]);
-		if ($this->request->is(['patch', 'post', 'put'])) {
-			$city = $this->Cities->patchEntity($city, $this->request->data);
-			if ($this->Cities->save($city)) {
-				$this->Flash->success('The city has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The city could not be saved. Please, try again.');
+		if ($this->request->params['_ext']) {
+			$city = $this->Cities->get($id);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$city = $this->Cities->patchEntity($city, $this->request->data);
+				if ($this->Cities->save($city, ['validate' => false])) {
+					$message = 'Saved';
+				} else {
+					$message = 'Error';
+				}
 			}
+			$this->set([
+				'city' => $city,
+				'message' => $message,
+				'data' => $this->request->data,
+				'_serialize' => ['message','city','data']
+			]);
 		}
-		$creators = $this->Cities->Creators->find('list');
-		$modifiers = $this->Cities->Modifiers->find('list');
-		$countries = $this->Cities->Countries->find('list');
-		$this->set(compact('city', 'creators', 'modifiers', 'countries'));
+		else {
+			$city = $this->Cities->get($id, [
+				'contain' => []
+			]);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$city = $this->Cities->patchEntity($city, $this->request->data);
+				if ($this->Cities->save($city)) {
+					$this->Flash->success('The city has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The city could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Cities->Creators->find('list');
+			$modifiers = $this->Cities->Modifiers->find('list');
+			$this->set(compact('city', 'creators', 'modifiers'));
+		}
 	}
 
 /**
@@ -105,13 +148,26 @@ class CitiesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function delete($id = null) {
-		$city = $this->Cities->get($id);
-		$this->request->allowMethod(['post', 'delete']);
-		if ($this->Cities->delete($city)) {
-			$this->Flash->success('The city has been deleted.');
-		} else {
-			$this->Flash->error('The city could not be deleted. Please, try again.');
+		if ($this->request->params['_ext']) {
+			$city = $this->Cities->get($id);
+			$message = 'Deleted';
+			if (!$this->Cities->delete($city)) {
+				$message = 'Error';
+			}
+			$this->set([
+				'message' => $message,
+				'_serialize' => ['message']
+			]);
 		}
-		return $this->redirect(['action' => 'index']);
+		else {		
+			$city = $this->Cities->get($id);
+			$this->request->allowMethod(['post', 'delete']);
+			if ($this->Cities->delete($city)) {
+				$this->Flash->success('The city has been deleted.');
+			} else {
+				$this->Flash->error('The city could not be deleted. Please, try again.');
+			}
+			return $this->redirect(['action' => 'index']);
+		}
 	}
 }
